refactor(modulith): close users — extract shared media module + expose NudgeUserDetailsService
Closing users surfaced two more OPEN-masked edges beyond users<->businesses:
- users -> servicesoffered (avatar cleanup): extracted the PendingMediaDeletion queue
  (entity + repository) into a new shared 'media' module; servicesoffered keeps the
  MediaCleanupJob/MediaService drainer that consumes the shared queue.
- config -> users.services.helpers.NudgeUserDetailsService (non-exposed): moved it up to the
  exposed users.services package (it's a service, not a helper).

users is now CLOSED with named interfaces {entities, repositories, services, spi, events}.
Every application module is CLOSED; config/utils/media are shared. verify() green; 291 tests.
