+++ src/main/java/com/mudhut/nudge/media/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.ApplicationModule(type = org.springframework.modulith.ApplicationModule.Type.OPEN)
+package com.mudhut.nudge.media;
