+++ src/main/java/com/mudhut/nudge/users/entities/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("entities")
+package com.mudhut.nudge.users.entities;
