+++ src/main/java/com/mudhut/nudge/users/events/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("events")
+package com.mudhut.nudge.users.events;
