+++ src/main/java/com/mudhut/nudge/users/package-info.java
@@ -1,2 +1,4 @@
-@org.springframework.modulith.ApplicationModule(type = org.springframework.modulith.ApplicationModule.Type.OPEN)
+// CLOSED module: only the named interfaces (entities, repositories, services, spi, events) are
+// part of the API other modules may reference; everything else is internal.
+@org.springframework.modulith.ApplicationModule
 package com.mudhut.nudge.users;
