+++ src/main/java/com/mudhut/nudge/users/repositories/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("repositories")
+package com.mudhut.nudge.users.repositories;
