+++ src/main/java/com/mudhut/nudge/users/services/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("services")
+package com.mudhut.nudge.users.services;
