+++ src/main/java/com/mudhut/nudge/users/spi/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("spi")
+package com.mudhut.nudge.users.spi;
