+++ src/main/kotlin/com/mudhut/nudge/servicesoffered/services/ServiceAddonService.kt
@@ -2,14 +2,14 @@ package com.mudhut.nudge.servicesoffered.services
 
 import com.mudhut.nudge.businesses.entities.BusinessRole
 import com.mudhut.nudge.businesses.services.BusinessService
-import com.mudhut.nudge.servicesoffered.entities.PendingMediaDeletion
+import com.mudhut.nudge.media.PendingMediaDeletion
 import com.mudhut.nudge.servicesoffered.entities.ServiceAddon
 import com.mudhut.nudge.servicesoffered.entities.ServiceOffered
 import com.mudhut.nudge.servicesoffered.models.CreateServiceAddonRequest
 import com.mudhut.nudge.servicesoffered.models.ReorderAddonsRequest
 import com.mudhut.nudge.servicesoffered.models.ServiceAddonResponse
 import com.mudhut.nudge.servicesoffered.models.UpdateServiceAddonRequest
-import com.mudhut.nudge.servicesoffered.repositories.PendingMediaDeletionRepository
+import com.mudhut.nudge.media.PendingMediaDeletionRepository
 import com.mudhut.nudge.servicesoffered.repositories.ServiceAddonRepository
 import com.mudhut.nudge.servicesoffered.repositories.ServiceOfferedRepository
 import com.mudhut.nudge.servicesoffered.events.ServiceAddonDeletedEvent
