+++ src/main/kotlin/com/mudhut/nudge/users/services/NudgeUserDetailsService.kt
@@ -1,4 +1,4 @@
-package com.mudhut.nudge.users.services.helpers
+package com.mudhut.nudge.users.services
 
 import com.mudhut.nudge.users.repositories.UserRepository
 import org.springframework.security.core.authority.SimpleGrantedAuthority
