+++ src/test/kotlin/com/mudhut/nudge/servicesoffered/controllers/ServiceOfferedControllerTest.kt
@@ -13,7 +13,7 @@ import com.mudhut.nudge.servicesoffered.models.MediaResponse
 import com.mudhut.nudge.servicesoffered.models.ServiceOfferedResponse
 import com.mudhut.nudge.servicesoffered.models.UpdateServiceOfferedRequest
 import com.mudhut.nudge.servicesoffered.services.ServicesOfferedService
-import com.mudhut.nudge.users.services.helpers.NudgeUserDetailsService
+import com.mudhut.nudge.users.services.NudgeUserDetailsService
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
