+++ src/test/kotlin/com/mudhut/nudge/servicesoffered/services/ServiceAddonServiceTest.kt
@@ -3,7 +3,7 @@ package com.mudhut.nudge.servicesoffered.services
 import com.mudhut.nudge.businesses.entities.Business
 import com.mudhut.nudge.businesses.entities.BusinessRole
 import com.mudhut.nudge.businesses.services.BusinessService
-import com.mudhut.nudge.servicesoffered.entities.PendingMediaDeletion
+import com.mudhut.nudge.media.PendingMediaDeletion
 import com.mudhut.nudge.servicesoffered.entities.PriceMode
 import com.mudhut.nudge.servicesoffered.entities.ServiceAddon
 import com.mudhut.nudge.servicesoffered.entities.ServiceOffered
@@ -11,7 +11,7 @@ import com.mudhut.nudge.servicesoffered.events.ServiceAddonDeletedEvent
 import com.mudhut.nudge.servicesoffered.models.CreateServiceAddonRequest
 import com.mudhut.nudge.servicesoffered.models.ReorderAddonsRequest
 import com.mudhut.nudge.servicesoffered.models.UpdateServiceAddonRequest
-import com.mudhut.nudge.servicesoffered.repositories.PendingMediaDeletionRepository
+import com.mudhut.nudge.media.PendingMediaDeletionRepository
 import com.mudhut.nudge.servicesoffered.repositories.ServiceAddonRepository
 import com.mudhut.nudge.servicesoffered.repositories.ServiceOfferedRepository
 import com.mudhut.nudge.utils.exceptions.ServiceAddonNotFoundException
